+++ storage/framework/core/actions/src/key-generate.ts
@@ -4,7 +4,6 @@ import { generateAppKey } from '@stacksjs/security'
 import { isFile } from '@stacksjs/storage'
 import { setEnvValue } from '@stacksjs/utils'
 
-if (!isFile('.env'))
-  await runCommand('cp .env.example .env', { cwd: projectPath() })
+if (!isFile('.env')) await runCommand('cp .env.example .env', { cwd: projectPath() })
 
 await setEnvValue('APP_KEY', generateAppKey())
