+++ storage/framework/core/actions/src/upgrade.ts
@@ -40,8 +40,7 @@ export async function downloadFrameworkUpdate(options: UpgradeOptions) {
   const tempFolderName = 'updates'
   const tempUpdatePath = projectPath(tempFolderName)
 
-  if (storage.doesFolderExist(tempUpdatePath))
-    await storage.deleteFolder(tempUpdatePath)
+  if (storage.doesFolderExist(tempUpdatePath)) await storage.deleteFolder(tempUpdatePath)
 
   log.info('Downloading framework updates...')
   await runCommand(`giget stacks ${tempFolderName}`, options)
